Examp2.4.3.1 from Duchateau, L. and Janssen, P. and Rowlands, G. J. (1998).Linear Mixed Models. An Introduction with applications in Veterinary Research. International Livestock Research Institute.

Examp2.4.3.1 is used for inspecting probability distribution and to define a plausible process through linear models and generalized linear models.

References

Duchateau, L. and Janssen, P. and Rowlands, G. J. (1998).Linear Mixed Models. An Introduction with applications in Veterinary Research. International Livestock Research Institute.

See also

Examples

#-------------------------------------------------------------
## Example 2.4.3.1 p-66
#-------------------------------------------------------------
 # PROC MIXED DATA=ex127;
 # CLASS sire;
 # MODEL ww=;
 # RANDOM sire/solution;
 # RUN;

library(lme4)
str(ex127)

#> 'data.frame':	43 obs. of  2 variables:
#>  $ sire: int  1 1 1 2 2 2 2 2 2 2 ...
#>  $ Ww  : num  11.1 10.8 11.3 15.9 13 12.6 14.6 13.6 12.5 15.7 ...

fm2.8 <-
  lme4::lmer(
         formula    = Ww~(1|sire)
       , data       = ex127
       , REML       = TRUE
       , control    = lmerControl()
       , start      = NULL
       , verbose    = 0L
    #  , subset
    #  , weights
    #  , na.action
    #  , offset
       , contrasts  = NULL
       , devFunOnly = FALSE
    #  , ...
       )
summary(fm2.8)

#> Linear mixed model fit by REML ['lmerMod']
#> Formula: Ww ~ (1 | sire)
#>    Data: ex127
#> 
#> REML criterion at convergence: 188.9
#> 
#> Scaled residuals: 
#>     Min      1Q  Median      3Q     Max 
#> -1.5024 -0.7106 -0.2259  0.8414  1.9459 
#> 
#> Random effects:
#>  Groups   Name        Variance Std.Dev.
#>  sire     (Intercept) 3.686    1.920   
#>  Residual             3.542    1.882   
#> Number of obs: 43, groups:  sire, 8
#> 
#> Fixed effects:
#>             Estimate Std. Error t value
#> (Intercept)  13.9686     0.7435   18.79

lme4::fixef(fm2.8)

#> (Intercept) 
#>    13.96863 

lme4::ranef(fm2.8)

#> $sire
#>   (Intercept)
#> 1  -2.1979037
#> 2  -0.0318874
#> 3  -1.7434719
#> 4   3.1587810
#> 5   0.8691909
#> 6  -0.8627974
#> 7   1.3556762
#> 8  -0.5475878
#>